servoshell: Move desktop main
